refactor: removes DDP stream support
Removes the native DDP stream support. This feature appears to be unused or replaced by a more modern approach, such as MongoDB Change Streams. The removal simplifies the codebase by eliminating related methods, classes, and message handling logic.
