--- validate_stream_subscription.js
@@ -1,48 +1,0 @@
-// Validation script for StreamSubscription refactoring
-
-import { Connection } from './packages/ddp-client/common/livedata_connection.js';
-
-// Test that StreamSubscription class is accessible from Connection instances
-const testStreamSubscriptionRefactoring = () => {
-  console.log('Testing StreamSubscription refactoring...');
-  
-  try {
-    // Create a mock stream for testing
-    const mockStream = {
-      send: () => {},
-      on: () => {},
-      status: () => ({ connected: false }),
-      reconnect: () => {},
-      disconnect: () => {}
-    };
-    
-    // Create a Connection instance
-    const conn = new Connection(mockStream, { retry: false });
-    
-    // Verify that StreamSubscription is accessible as a class property
-    if (typeof conn.StreamSubscription === 'function') {
-      console.log('✅ StreamSubscription is accessible as Connection.StreamSubscription');
-    } else {
-      console.log('❌ StreamSubscription is not accessible');
-      return false;
-    }
-    
-    // Verify that _streamSubscriptions is initialized
-    if (typeof conn._streamSubscriptions === 'object' && conn._streamSubscriptions !== null) {
-      console.log('✅ _streamSubscriptions is properly initialized');
-    } else {
-      console.log('❌ _streamSubscriptions is not properly initialized');
-      return false;
-    }
-    
-    console.log('✅ All StreamSubscription refactoring tests passed!');
-    return true;
-    
-  } catch (error) {
-    console.log('❌ Error during testing:', error.message);
-    return false;
-  }
-};
-
-// Run the test
-testStreamSubscriptionRefactoring();
